+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Added
 -----
 
 * Add support for more date format codes (days, weeks, months and years) in the `file` setting
+* Add new `round_entries` setting to customize entries duration rounding
 
 5.0 (2019-12-27)
 ================